Created API endpoint, pushed class out to separate file
