+++ Dockerfile/Dockerfile
@@ -1,0 +1,10 @@
+FROM python
+
+# # all commands start from this directory
+# WORKDIR /componentscraper
+#
+# # copy all files from this folder to working directory (ignores files in .dockerignore)
+# COPY . .
+
+# set the start command
+CMD [ "python3", "pricefinder.py" ]
