+++ README.md
@@ -5,3 +5,4 @@ Scraper to get prices for components for desired build
 
 Send notification to phone on low price/historic low
 add to gcp
+what happens if the request isnt 200?
