+++ flask/bin/python
@@ -1,0 +1,1 @@
+/usr/bin/python3
