+++ flask/bin/wheel-3.8
@@ -1,0 +1,8 @@
+#!/mnt/c/coding/Scraping/componentscraper/flask/bin/python
+# -*- coding: utf-8 -*-
+import re
+import sys
+from wheel.cli import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
